+++ src/ray_tracing_library/src/Vec3.cu
@@ -15,22 +15,22 @@ __host__ __device__ Vec3 &Vec3::operator+=(const Vec3 &v) {
     return *this;
 }
 
-__host__ __device__ Vec3 &Vec3::operator*=(const double t) {
+__host__ __device__ Vec3 &Vec3::operator*=(const float t) {
     _x[0] *= t;
     _x[1] *= t;
     _x[2] *= t;
     return *this;
 }
 
-__host__ __device__ Vec3 &Vec3::operator/=(const double t) {
+__host__ __device__ Vec3 &Vec3::operator/=(const float t) {
     return *this *= 1 / t;
 }
 
-__host__ __device__ double Vec3::length() const {
-    return sqrt(length_squared());
+__host__ __device__ float Vec3::length() const {
+    return sqrtf(length_squared());
 }
 
-__host__ __device__ double Vec3::length_squared() const {
+__host__ __device__ float Vec3::length_squared() const {
     return _x[0] * _x[0] + _x[1] * _x[1] + _x[2] * _x[2];
 }
 
@@ -51,19 +51,19 @@ __host__ __device__ Vec3 operator*(const Vec3 &u, const Vec3 &v) {
     return {u.x() * v.x(), u.y() * v.y(), u.z() * v.z()};
 }
 
-__host__ __device__ Vec3 operator*(double t, const Vec3 &v) {
+__host__ __device__ Vec3 operator*(float t, const Vec3 &v) {
     return {t * v.x(), t * v.y(), t * v.z()};
 }
 
-__host__ __device__ Vec3 operator*(const Vec3 &v, double t) {
+__host__ __device__ Vec3 operator*(const Vec3 &v, float t) {
     return t * v;
 }
 
-__host__ __device__ Vec3 operator/(Vec3 v, double t) {
-    return (1.0 / t) * v;
+__host__ __device__ Vec3 operator/(Vec3 v, float t) {
+    return (1.0f / t) * v;
 }
 
-__host__ __device__ double dot(const Vec3 &u, const Vec3 &v) {
+__host__ __device__ float dot(const Vec3 &u, const Vec3 &v) {
     return u.x() * v.x()
            + u.y() * v.y()
            + u.z() * v.z();
@@ -80,13 +80,13 @@ __host__ __device__ Vec3 normalize(Vec3 v) {
 }
 
 __host__ __device__ Vec3 reflect(const Vec3 &v, const Vec3 &n) {
-    return v - 2 * dot(v, n) * n;
+    return v - 2.0f * dot(v, n) * n;
 }
 
-__host__ __device__ Vec3 refract(const Vec3 &rayDirNormalized, const Vec3 &n, double refractionIdxRatio) {
-    double cosTheta = fmin(dot(-rayDirNormalized, n), 1.0);
+__host__ __device__ Vec3 refract(const Vec3 &rayDirNormalized, const Vec3 &n, float refractionIdxRatio) {
+    float cosTheta = fminf(dot(-rayDirNormalized, n), 1.0f);
     Vec3 r_out_perp = refractionIdxRatio * (rayDirNormalized + cosTheta * n);
-    Vec3 r_out_parallel = -sqrt(fabs(1.0 - r_out_perp.length_squared())) * n;
+    Vec3 r_out_parallel = -sqrtf(fabsf(1.0f - r_out_perp.length_squared())) * n;
     return r_out_perp + r_out_parallel;
 }
 
@@ -103,17 +103,17 @@ __device__ Vec3 randomVec0to1(curandState *randState) {
 }
 
 __host__ Vec3 randomUnitVec(int &randState) {
-    return normalize({2 * randomDouble(randState) - 1, 2 * randomDouble(randState) - 1, 2 * randomDouble(randState) - 1});
+    return normalize({2.0f * randomDouble(randState) - 1.0f, 2.0f * randomDouble(randState) - 1.0f, 2.0f * randomDouble(randState) - 1.0f});
 }
 
 __device__ Vec3 randomUnitVec(curandState *randState) {
-    return normalize({2 * randomFloatCuda(randState) - 1, 2 * randomFloatCuda(randState) - 1, 2 * randomFloatCuda(randState) - 1});
+    return normalize({2.0f * randomFloatCuda(randState) - 1.0f, 2.0f * randomFloatCuda(randState) - 1.0f, 2.0f * randomFloatCuda(randState) - 1.0f});
 }
 
 __host__ Vec3 randomVecInUnitDisk(int &randState) {
     while (true) {
         auto p = randomUnitVec(randState);
-        if (1 <= p.length_squared()) {
+        if (1.0f <= p.length_squared()) {
             continue;
         }
         return p;
@@ -123,7 +123,7 @@ __host__ Vec3 randomVecInUnitDisk(int &randState) {
 __device__ Vec3 randomVecInUnitDisk(curandState *randState) {
     while (true) {
         auto p = randomUnitVec(randState);
-        if (1 <= p.length_squared()) {
+        if (1.0f <= p.length_squared()) {
             continue;
         }
         return p;
@@ -148,21 +148,21 @@ __device__ Vec3 randomVecInUnitDisk(curandState *randState) {
 //}
 
 __host__ __device__ bool isZeroVec(const Vec3 &v) {
-    return fcmp(v.x(), 0) && fcmp(v.y(), 0) && fcmp(v.z(), 0);
+    return fcmp(v.x(), 0.0f) && fcmp(v.y(), 0.0f) && fcmp(v.z(), 0.0f);
 }
 
 __host__ __device__
-Vec3 alphaBlending(Vec3 v1, Vec3 v2, double alpha) {
+Vec3 alphaBlending(Vec3 v1, Vec3 v2, float alpha) {
     return v1 * (1 - alpha) + v2 * alpha;
 }
 
 __host__ __device__
 Color gammaCorrection(const Color &color) {
-    return {std::sqrt(color.x()), std::sqrt(color.y()), std::sqrt(color.z())};
+    return {sqrtf(color.x()), sqrtf(color.y()), sqrtf(color.z())};
 }
 
 __host__ __device__
-Vec3 clamp(const Vec3 &toClamp, double low, double high) {
+Vec3 clamp(const Vec3 &toClamp, float low, float high) {
     return {clamp(toClamp.x(), low, high),
             clamp(toClamp.y(), low, high),
             clamp(toClamp.z(), low, high)};
