+++ src/ray_tracing_library/src/my_math.cpp
@@ -4,12 +4,12 @@
 
 #include "my_math.h"
 
-double randomDouble(int &randState) {
-    static std::uniform_real_distribution<double> distribution(0.0, 1.0);
+float randomDouble(int &randState) {
+    static std::uniform_real_distribution<float> distribution(0.0, 1.0);
     static std::minstd_rand generator(1);
     return distribution(generator);
 }
 
-double randomDouble(int &randState, double from, double to) {
+float randomDouble(int &randState, float from, float to) {
     return randomDouble(randState) * (from - to) + from;
 }
